bottom navbar, created an validation for roles, fixing some major bugs where inference doesn't disposed and causing a leaking memory, creating an explore page UI.
