+++ assets/labelmap.txt
@@ -1,91 +1,91 @@
 ???
-person
-bicycle
-car
-motorcycle
-airplane
-bus
-train
-truck
-boat
-traffic light
-fire hydrant
+Person
+Bicycle
+Car
+Motorcycle
+Airplane
+Bus
+Train
+Truck
+Boat
+Traffic Light
+Fire Hydrant
 ???
-stop sign
-parking meter
-bench
-bird
-cat
-dog
-horse
-sheep
-cow
-elephant
-bear
-zebra
-giraffe
+Stop Sign
+Parking Meter
+Bench
+Bird
+Cat
+Dog
+Horse
+Sheep
+Cow
+Elephant
+Bear
+Zebra
+Giraffe
 ???
-backpack
-umbrella
-???
-???
-handbag
-tie
-suitcase
-frisbee
-skis
-snowboard
-sports ball
-kite
-baseball bat
-baseball glove
-skateboard
-surfboard
-tennis racket
-bottle
-???
-wine glass
-cup
-fork
-knife
-spoon
-bowl
-banana
-apple
-sandwich
-orange
-broccoli
-carrot
-hot dog
-pizza
-donut
-cake
-chair
-couch
-potted plant
-bed
-???
-dining table
+Backpack
+Umbrella
 ???
 ???
-toilet
+Handbag
+Tie
+Suitcase
+Frisbee
+Skis
+Snowboard
+Sports Ball
+Kite
+Baseball Bat
+Baseball Glove
+Skateboard
+Surfboard
+Tennis Racket
+Bottle
 ???
-tv
-laptop
-mouse
-remote
-keyboard
-cell phone
-microwave
-oven
-toaster
-sink
-refrigerator
+Wine Glass
+Cup
+Fork
+Lnife
+Spoon
+Bowl
+Banana
+Apple
+Sandwich
+Orange
+Broccoli
+Carrot
+Hot dog
+Pizza
+Donut
+Cake
+Chair
+Couch
+Potted plant
+Bed
 ???
-book
-clock
-vase
-scissors
-teddy bear
-hair drier
-toothbrush
+Dining table
+???
+???
+Toilet
+???
+Television
+Laptop
+Mouse
+Remote
+Keyboard
+Phone
+Microwave
+Oven
+Toaster
+Sink
+Refrigerator
+???
+Book
+Clock
+Vase
+Scissors
+Teddy bear
+Hair drier
+Toothbrush
