--- assets/labels.txt
@@ -1,2 +1,0 @@
-0 Uang 2 Ribu
-1 Uang 10 Ribu
