+++ lib/constant/constant_builder.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+const Color white = Colors.white;
+const Color lightGrey = Color(0xffAFAFAF);
+const Color darkGrey = Color(0xff666666);
+const Color appOrange = Color(0xffFF4000);
+const Color green = Color(0xff74E087);
+const Color appYellow = Color(0xffFFC000);
+const Color fontColor = Color(0xff3F3F3F);
+const Color selectedNavBar = Color(0xffFFC000);
+
+
+const appLogo = 'assets/image/seekers_logo/app_logo.png';
+const splashLogo = 'assets/image/seekers_logo/splash_logo.png';
+
+const explore = 'assets/image/icon/explore.png';
+const game = 'assets/image/icon/game.png';
+const history = 'assets/image/icon/history.png';
+const home = 'assets/image/icon/home.png';
+const mic = 'assets/image/icon/microphone.png';
+const profile = 'assets/image/icon/profile.png';
+const save = 'assets/image/icon/save.png';
+const hear = 'assets/image/icon/sound.png';
+const hearDark = 'assets/image/icon/sound_dark.png';
+
+const completedTask = 'assets/image/illustration/all_good.png';
+const articleBg = 'assets/image/illustration/article_bg.png';
+const clickMe = 'assets/image/illustration/click_me.png';
+const profilePic = 'assets/image/illustration/profile.png';
+const selectRoles = 'assets/image/illustration/select_roles.png';
+const walkingMan = 'assets/image/illustration/walking_man.png';
