--- lib/stats.dart
@@ -1,25 +1,0 @@
-class Stats {
-  /// Total time taken in the isolate where the inference runs
-  int? totalPredictTime;
-
-  /// [totalPredictTime] + communication overhead time
-  /// between main isolate and another isolate
-  int? totalElapsedTime;
-
-  /// Time for which inference runs
-  int? inferenceTime;
-
-  /// Time taken to pre-process the image
-  int? preProcessingTime;
-
-  Stats(
-      {this.totalPredictTime,
-      this.totalElapsedTime,
-      this.inferenceTime,
-      this.preProcessingTime});
-
-  @override
-  String toString() {
-    return 'Stats{totalPredictTime: $totalPredictTime, totalElapsedTime: $totalElapsedTime, inferenceTime: $inferenceTime, preProcessingTime: $preProcessingTime}';
-  }
-}
