+++ lib/tflite/image_utils.dart
@@ -1,0 +1,80 @@
+import 'dart:io';
+
+import 'package:camera/camera.dart';
+import 'package:image/image.dart' as imageLib;
+import 'package:path_provider/path_provider.dart';
+
+/// ImageUtils
+class ImageUtils {
+  /// Converts a [CameraImage] in YUV420 format to [imageLib.Image] in RGB format
+  static imageLib.Image? convertCameraImage(CameraImage cameraImage) {
+    if (cameraImage.format.group == ImageFormatGroup.yuv420) {
+      return convertYUV420ToImage(cameraImage);
+    } else if (cameraImage.format.group == ImageFormatGroup.bgra8888) {
+      return convertBGRA8888ToImage(cameraImage);
+    } else {
+      return null;
+    }
+  }
+
+  /// Converts a [CameraImage] in BGRA888 format to [imageLib.Image] in RGB format
+  static imageLib.Image convertBGRA8888ToImage(CameraImage cameraImage) {
+    imageLib.Image img = imageLib.Image.fromBytes(cameraImage.planes[0].width ?? 300,
+        cameraImage.planes[0].height ?? 400, cameraImage.planes[0].bytes,
+        format: imageLib.Format.bgra);
+    return img;
+  }
+
+  /// Converts a [CameraImage] in YUV420 format to [imageLib.Image] in RGB format
+  static imageLib.Image convertYUV420ToImage(CameraImage cameraImage) {
+    final int width = cameraImage.width;
+    final int height = cameraImage.height;
+
+    final int uvRowStride = cameraImage.planes[1].bytesPerRow;
+    final int uvPixelStride = cameraImage.planes[1].bytesPerPixel ?? 100;
+
+    final image = imageLib.Image(width, height);
+
+    for (int w = 0; w < width; w++) {
+      for (int h = 0; h < height; h++) {
+        final int uvIndex =
+            uvPixelStride * (w / 2).floor() + uvRowStride * (h / 2).floor();
+        final int index = h * width + w;
+
+        final y = cameraImage.planes[0].bytes[index];
+        final u = cameraImage.planes[1].bytes[uvIndex];
+        final v = cameraImage.planes[2].bytes[uvIndex];
+
+        image.data[index] = ImageUtils.yuv2rgb(y, u, v);
+      }
+    }
+    return image;
+  }
+
+  /// Convert a single YUV pixel to RGB
+  static int yuv2rgb(int y, int u, int v) {
+    // Convert yuv pixel to rgb
+    int r = (y + v * 1436 / 1024 - 179).round();
+    int g = (y - u * 46549 / 131072 + 44 - v * 93604 / 131072 + 91).round();
+    int b = (y + u * 1814 / 1024 - 227).round();
+
+    // Clipping RGB values to be inside boundaries [ 0 , 255 ]
+    r = r.clamp(0, 255);
+    g = g.clamp(0, 255);
+    b = b.clamp(0, 255);
+
+    return 0xff000000 |
+        ((b << 16) & 0xff0000) |
+        ((g << 8) & 0xff00) |
+        (r & 0xff);
+  }
+
+  static void saveImage(imageLib.Image image, [int i = 0]) async {
+    List<int> jpeg = imageLib.JpegEncoder().encodeImage(image);
+    final appDir = await getTemporaryDirectory();
+    final appPath = appDir.path;
+    final fileOnDevice = File('$appPath/out$i.jpg');
+    await fileOnDevice.writeAsBytes(jpeg, flush: true);
+    print('Saved $appPath/out$i.jpg');
+  }
+}
