+++ lib/view/history_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class HistoryPage extends StatelessWidget {
+  const HistoryPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('History page'),
+    );
+  }
+}
