+++ lib/view/impaired/game_impaired_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class GameImpaired extends StatelessWidget {
+  const GameImpaired({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('game impaired page'),
+    );
+  }
+}
