+++ lib/view/peer/game_peer_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class GamePeer extends StatelessWidget {
+  const GamePeer({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('game peer page'),
+    );
+  }
+}
