+++ lib/view/peer/home_peer_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class HomePeerPage extends StatelessWidget {
+  const HomePeerPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('home peer page'),
+    );
+  }
+}
