+++ lib/view/profile_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class ProfilePage extends StatelessWidget {
+  const ProfilePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text('Profile page'),
+    );
+  }
+}
